install-nsight-systems.sh: run conditionally (#1496)

+++ .github/container/install-nsight-systems.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 set -exuo pipefail
 
+version=$1
+if [[ -z "${version}" ]]; then
+  echo "Usage: $0 major.minor.patch"
+  exit 1
+fi
+
 # Remove the symlink that makes `nsys` refer to the CUDA-bundled version:
 rm /usr/local/cuda/bin/nsys
 
@@ -15,7 +21,7 @@ export DEBIAN_FRONTEND=noninteractive
 export TZ=America/Los_Angeles
 
 apt-get update
-apt-get install -y nsight-systems-cli-2025.3.1
+apt-get install -y nsight-systems-cli-${version}
 apt-get clean
 
 rm -rf /var/lib/apt/lists/*
